restructured code
