+++ src/components/index.js
@@ -37,6 +37,6 @@ export { BottomTabBar } from "./BottomTabBar";
 export { AnimatedImage } from "./AnimatedImage";
 export { NotificationPreferenceList } from "./NotificationPreferenceList";
 export { OrganizationItem } from "./OrganizationItem";
-export { ChatInput } from "./ChatInput";
+export { ChatInput } from "./ChatInput/ChatInput";
 export { Loader } from "./Loader";
 export { InputEmailChip } from "./InputEmailChip";
